add key in @Headers

+++ src/order/order.controller.ts
@@ -3,6 +3,7 @@ import {
   Controller,
   Delete,
   Get,
+  Headers,
   Param,
   Post,
   Put,
@@ -10,12 +11,14 @@ import {
 import { OrderDto } from './DTO/order.dto';
 import { OrderService } from './order.service';
 import { ProductService } from '../product/product.service';
+import { UserService } from '../user/user.service';
 
 @Controller('order')
 export class OrderController {
   constructor(
     private readonly orderService: OrderService,
     private readonly productService: ProductService,
+    private readonly userService: UserService,
   ) {}
   /**
    * Create a new order
@@ -32,18 +35,33 @@ export class OrderController {
    * @returns A list of all orders
    */
   @Get()
-  async getAllOrders() {
-    return await this.orderService.getAllOrders();
+  async getAllOrders(
+    @Headers('key')
+    key: string,
+  ) {
+    if (await this.userService.validateToken(key)) {
+      return await this.orderService.getAllOrders();
+    }
+    return false;
   }
 
   /**
    * Remove an order by ID
+   * @param key
    * @param id - The ID of the order to remove
    * @returns A boolean indicating if the order was successfully removed
    */
   @Delete(':id')
-  async removeOrder(@Param('id') id: number) {
-    return await this.orderService.removeOrder(id);
+  async removeOrder(
+    @Param('id')
+    @Headers('key')
+    key: string,
+    id: number,
+  ) {
+    if (await this.userService.validateToken(key)) {
+      return await this.orderService.removeOrder(id);
+    }
+    return false;
   }
 
   /**
@@ -52,12 +70,20 @@ export class OrderController {
    * Update an order by ID
    * @param id - The ID of the order to update
    * @param updateOrderDto - The data to update the order with
+   * @param key
    * @returns The updated order
    */
   @Put(':id')
-  async updateOrder(@Param('id') id: number, @Body() updateOrderDto: OrderDto) {
+  async updateOrder(
+    @Param('id') id: number,
+    @Body() updateOrderDto: OrderDto,
+    @Headers('key')
+    key: string,
+  ) {
     return false;
-    const ids = updateOrderDto.products.map((product) => product.id);
-    return await this.productService.findByIds(ids);
+    if (await this.userService.validateToken(key)) {
+      return await this.orderService.updateOrder(id, updateOrderDto);
+    }
+    return false;
   }
 }
